fix: change duration field type in AdPhrase model from int to float for better precision

+++ backend/models/media.py
@@ -24,7 +24,7 @@ class AdPhrase(BaseModel):
     voice_speaker: str
     voice_sex: str
     type: str
-    duration: int = -1
+    duration: float = -1
 
 class AddAdPhraseRequest(BaseModel):
     user_id: str
